Move subformats into separate files
We’re going to be adding more formats so it probably makes sense to
break things up a bit to keep things manageable.
